helper for underlyingDataSummary
